pyproject.toml
[tool.poetry]
name = "psd2alltheway"
version = "0.1.0"
description = ""
authors = ["Daniel Steman <daniel-steman@live.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
